+++ public/mix-manifest.json
@@ -1,4 +1,4 @@
 {
-    "/js/app.js": "/js/app.js?id=99156118b1a486332889",
+    "/js/app.js": "/js/app.js?id=7167e25d29699d3942b0",
     "/css/app.css": "/css/app.css?id=b65288437bcadd420bea"
 }
