+++ resources/js/components/task/List.vue
